fix: add a `@deprecated` tag to the `mswDecorator` (#178)
* Add an `@deprecated` JSDoc tag to the mswDecorator.

Previously, this was only communicated to the user via the browser console, and so was easy to miss. I only noticed it because of reading the changelog.

It'll be much more likely for users to notice this deprecation if their editor tells them that it's deprecated, so this will hopefully help ease migration.

* Update packages/msw-addon/src/decorator.ts

Co-authored-by: Artem Zakharchenko <kettanaito@gmail.com>

---------

Co-authored-by: Artem Zakharchenko <kettanaito@gmail.com>